Updates to allow for easy access/serialization of virtual attributes
Changed ActiveRecord\Model so that the __get() method will look for
a method with the same name as the property requested, in addition
to looking for the get_property_name() getter method.

Changed ActiveRecord\Serialization so that the check_methods() method
will look for a getter method in the form of "get_method_name" if it
can't find a method "method_name" in the class. This allows custom
attributes to be easily included in serialization.